analytics/requirements.txt
psycopg2-binary
pika
